Refactor the codebase into a more organized structure with `agent`, `tool`, and `task` directories.
- **Agent:** The `agent` directory contains the logic for interacting with the Google Gemini API.
- **Tool:** The `tool` directory contains utilities for parsing EPUB files (using Tika) and CSV files.
- **Task:** The `task` directory orchestrates the workflow, calling the agent and tools to process the book data.
- **Models:** Pydantic models were moved to a dedicated `src/models` directory.

The main `main.py` script has been updated to use the new refactored structure. This change improves code organization and maintainability.
